backend/requirements.txt
django
whitenoise
psycopg2-binary==2.8.5
gunicorn
psycopg2
dj-database-url==0.5.0
djangorestframework